Refactor common test code into reusable files
