+++ tests/client/setup.js
@@ -1,0 +1,5 @@
+import { mockConnection } from '../../client/js/connection.js';
+
+mockConnection();
+
+document.body.insertAdjacentHTML('beforeend', '<div id="roomArea"> <div id="room"> <div id="topSurface" class="surface"></div> </div></div> <div id="debugButtonOverlay"><pre id="debugButtonOutput"></pre></div> <div id="enlarged"></div>');
